@@ -1,7 +1,5 @@
-# frozen_string_literal: true
-
-# Load the rails application
+# Load the Rails application.
 require File.expand_path('../application', __FILE__)
 
-# Initialize the rails application
+# Initialize the Rails application.
 Contact::Application.initialize!
